vibez-intelligence/requirements.txt
Flask
pymongo
openai
python-dotenv
flask-cors
